Renames package to discordplays
